requirements.txt
# Function dependencies, for example:
# package>=version
slack-bolt==1.20.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-cloud-aiplatform==1.53.0